# servicios/transparencia/src/main/resources/openapi/transparencia.yaml
...
openapi: 3.1.0
info:
  title: transparencia
  version: 1.0.0
  description: >-
    Servicio transparencia. Un caso de uso = una operacion, con el codigo CU en el
    operationId. Los importes viajan como CADENA decimal, nunca como number.

    Este servicio existe para que nadie tenga que creernos. Por eso tiene las dos
    unicas familias de rutas sin sesion del sistema: /publico y /verificar.
servers:
  - url: /api/v1
paths:
...
  /reputacion/eventos:
    post:
      operationId: registrarEventoReputacion    # CU-70
      summary: Registrar un evento de reputacion
      description: >-
        Un evento por hecho (R-REP-01). El impacto sale del catalogo del modelo, no de
        constantes: sin regla no se puntua, porque inventarle un impacto a un tipo de
        evento es decidir a mano cuanto vale la conducta de alguien.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaEventoReputacion' }
      responses:
        '201':
          description: Evento registrado, con su impacto y el factor que mueve
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaEventoReputacion' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/eventos/{eventoId}/compensacion:
    post:
      operationId: compensarEventoReputacion    # CU-70
      summary: Compensar un evento que sumo o resto puntos
      description: >-
        No se borra el original ni se lo edita: evento_reputacion es append-only. Entra
        una fila nueva con el impacto exacto invertido y `revertidoPorId` apuntando al
        original.
      parameters:
        - name: eventoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCompensacion' }
      responses:
        '201':
          description: Evento compensatorio registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaEventoReputacion' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/puntajes:
    post:
      operationId: recalcularPuntaje            # CU-71
      summary: Recalcular el puntaje de reputacion
      description: >-
        Un solo puntaje vigente por usuario (R-REP-02) y el total es la suma de sus
        componentes (R-REP-03). Con menos eventos que el minimo del modelo el nivel es
        SIN_HISTORIAL: no es un castigo, es decir la verdad.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRecalculo' }
      responses:
        '200':
          description: Puntaje vigente, abierto factor por factor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaPuntaje' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/snapshots:
    post:
      operationId: tomarSnapshotReputacion      # CU-71
      summary: Congelar una foto del puntaje
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaSnapshot' }
      responses:
        '201':
          description: Foto tomada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaSnapshot' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/bloques:
    post:
      operationId: sellarBloque                 # CU-72
      summary: Sellar el bloque de transparencia
      description: >-
        Con excepciones de conciliacion abiertas NO se sella (R-BIL-12). Un hecho ya
        sellado no se reescribe: la correccion entra al bloque siguiente como
        movimiento compensatorio y los dos quedan visibles.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaSellado' }
      responses:
        '201':
          description: Bloque sellado y encadenado al anterior
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaSellado' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
...
  /reputacion/insignias/evaluacion:
    post:
      operationId: evaluarInsignias             # CU-74
      summary: Evaluar y otorgar las insignias que un hecho desbloqueo
      description: >-
        NO hay otorgamiento manual. Se gana contra los datos o no se gana: una insignia
        que alguien puede conceder a dedo no dice nada de quien la lleva.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaEvaluacionInsignias' }
      responses:
        '200':
          description: Lo otorgado, con su motivo legible
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SalidaInsignia' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/insignias/{otorgadaId}/revocacion:
    post:
      operationId: revocarInsignia              # CU-74
      summary: Revocar una insignia, sin borrarla
      description: >-
        Revocar no borra (R-REP-05). Queda la fila, la fecha y el motivo: borrarla
        dejaria a la persona sin poder saber por que perdio algo que tenia.
      parameters:
        - name: otorgadaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRevocacionInsignia' }
      responses:
        '200':
          description: Insignia revocada, con su motivo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaInsignia' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/certificados:
    post:
      operationId: emitirCertificado            # CU-75
      summary: Emitir un certificado de reputacion verificable
      description: >-
        El titular elige que incluir campo por campo, y lo que no eligio no entra ni al
        hash. Solo el titular emite el suyo (R-SEG-03).
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCertificado' }
      responses:
        '201':
          description: Certificado emitido, con su URL publica y su vencimiento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaCertificado' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/certificados/{codigo}/revocacion:
    post:
      operationId: revocarCertificado           # CU-75
      summary: Revocar un certificado
      description: >-
        Lo usa tambien el derecho de supresion (CU-07): al suprimir, la URL deja de
        resolver.
      parameters:
        - name: codigo
          in: path
          required: true
          schema: { type: string, maxLength: 40 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRevocacionCertificado' }
      responses:
        '204': { description: Certificado revocado }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
...
  /reputacion/resenas:
    post:
      operationId: publicarResena               # CU-76
      summary: Resenar a un participante
      description: >-
        Solo resena quien convivio y una vez por autor, evaluado, grupo y dimension
        (R-REP-06). Se resena al cerrar el ciclo, no durante: resenar a quien todavia
        te debe plata convierte la resena en una herramienta de presion.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaResena' }
      responses:
        '201':
          description: Resena creada, PENDIENTE de moderacion
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaResena' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/resenas/{resenaId}/moderacion:
    post:
      operationId: moderarResena                # CU-76
      summary: Moderar una resena
      description: >-
        Nada se publica sin moderar. Lo que trae datos personales se retiene, no se
        recorta: recortar deja el dato en la base y la sospecha en el aire.
      parameters:
        - name: resenaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaModeracion' }
      responses:
        '200':
          description: Resena moderada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaResena' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/riesgo/evaluacion:
    post:
      operationId: evaluarRiesgo                # CU-97
      summary: Anticipar el riesgo con alertas tempranas
      description: >-
        Sin historial el nivel es SIN_DATOS, nunca riesgo alto. Al participante nunca
        se le muestra un puntaje: el mensaje habla de hechos concretos, con numeros, y
        de que puede hacer.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaEvaluacionRiesgo' }
      responses:
        '200':
          description: Alertas abiertas, nivel de riesgo y metricas en alerta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaEvaluacionRiesgo' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/riesgo/alertas/{alertaId}/cierre:
    post:
      operationId: cerrarAlertaRiesgo           # CU-97
      summary: Cerrar una alerta con su desenlace
      description: >-
        Cerrar exige desenlace: una alerta que se cierra sin decir si era cierta no
        deja calibrar nada, y el modelo se queda con el mismo error para siempre.
      parameters:
        - name: alertaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCierreAlerta' }
      responses:
        '204': { description: Alerta cerrada con su desenlace }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /reputacion/{usuarioId}/puntaje:
    get:
      operationId: consultarPuntaje
      summary: El puntaje de reputacion de alguien
      description: >-
        Lo preguntan CU-64 y CU-68 de `grupos`: aceptar a alguien o dejarle tomar un
        cupo depende de su historial, y el historial es de este servicio.

        Quien todavia no tiene puntaje NO tiene cero. Cero seria el peor puntaje
        posible y quien recien llega no hizo nada malo: se responde SIN_HISTORIAL y
        quien pregunta decide. Confundir «no se sabe» con «malo» le cierra la puerta a
        todo el que empieza.
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: El puntaje, o la constancia de que no hay historial
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PuntajeDelUsuario' }
components:
  parameters:
    ClaveIdempotencia:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Toda operacion con efecto la exige. La red duplica; reintentar tiene que ser
        seguro, y la clave se valida ANTES de escribir (invariante 7).
      schema: { type: string, format: uuid }
  responses:
    ReglaDeNegocio:
      description: Una regla del dominio lo impide. El codigo AP-CUnn-nn dice cual.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
  schemas:
...
    Hash:
      type: string
      minLength: 64
      maxLength: 64
      description: SHA-256 en hexadecimal minuscula.
...
    EntradaEventoReputacion:
      type: object
      additionalProperties: false
      required: [usuarioId, tipo, referenciaTipo, descripcion, ocurridoEn]
      properties:
        usuarioId:      { type: string, format: uuid }
        grupoId:        { type: [string, 'null'], format: uuid }
        participanteId: { type: [string, 'null'], format: uuid }
        tipo:
          type: string
          description: Catalogo cerrado de ck_evento_reputacion_tipo.
          enum: [ABANDONO_DE_GRUPO, APORTE_ANTICIPADO, APORTE_CON_RETRASO_GRAVE,
                 APORTE_CON_RETRASO_LEVE, APORTE_NO_REALIZADO, APORTE_PUNTUAL,
                 AVAL_EJECUTADO, COBERTURA_CONSUMIDA, CONFIRMACION_ENTREGA_A_TIEMPO,
                 DEUDA_CASTIGADA, DEUDA_RECUPERADA, EXPULSION, FRAUDE_CONFIRMADO,
                 GRUPO_COMPLETADO, KYC_COMPLETADO, RESENA_NEGATIVA, RESENA_POSITIVA,
                 SANCION_APLICADA, SANCION_REVOCADA]
        referenciaTipo:
          type: string
          enum: [COBERTURA_INCUMPLIMIENTO, ENTREGA_FONDO, GRUPO, OBLIGACION_APORTE,
                 PARTICIPANTE, REGISTRO_INCUMPLIMIENTO, RESENA_PARTICIPANTE, TRASPASO_CUPO]
        referenciaId:  { type: [string, 'null'], format: uuid }
        descripcion:   { type: string, maxLength: 200 }
        esReversible:  { type: boolean, default: false }
        ocurridoEn:    { type: string, format: date-time }
    EntradaCompensacion:
      type: object
      additionalProperties: false
      required: [descripcion]
      properties:
        descripcion: { type: string, maxLength: 200 }
    SalidaEventoReputacion:
      type: object
      additionalProperties: false
      required: [eventoId, impacto, esNuevo]
      properties:
        eventoId:       { type: string, format: uuid }
        impacto:        { $ref: '#/components/schemas/Decimal' }
        factorAfectado: { type: [string, 'null'] }
        esNuevo:
          type: boolean
          description: false cuando el hecho ya estaba puntuado (R-REP-01). El reintento es inocuo.

    # ------------------------------------------------------------------ CU-71
...
    EntradaRecalculo:
      type: object
      additionalProperties: false
      required: [usuarioId, mediciones, indicadores]
      properties:
        usuarioId:  { type: string, format: uuid }
        mediciones: { type: array, items: { $ref: '#/components/schemas/Medicion' } }
        medicionesAnteriores:
          type: array
          items: { $ref: '#/components/schemas/Medicion' }
          description: Para la tendencia. Alguien con 60 subiendo esta mejor que alguien con 65 cayendo.
        indicadores: { $ref: '#/components/schemas/Indicadores' }
...
    SalidaPuntaje:
      type: object
      additionalProperties: false
      required: [puntajeId, puntaje, nivelDeConfianza, componentes, eventosConsiderados]
      properties:
        puntajeId:        { type: string, format: uuid }
        puntaje:          { $ref: '#/components/schemas/Decimal' }
        nivelDeConfianza:
          type: string
          enum: [SIN_HISTORIAL, RESTRINGIDO, EN_OBSERVACION, BASICO, CONFIABLE,
                 MUY_CONFIABLE, REFERENTE]
        componentes:
          type: array
          items: { $ref: '#/components/schemas/Componente' }
          description: Su suma es el total (R-REP-03). Es lo que hace que el numero se pueda apelar.
        puntajeAnterior:     { type: [string, 'null'] }
        eventosConsiderados: { type: integer, minimum: 0 }
    EntradaSnapshot:
      type: object
      additionalProperties: false
      required: [usuarioId, motivo]
      properties:
        usuarioId: { type: string, format: uuid }
        motivo:
          type: string
          enum: [AUDITORIA, CIERRE_DE_GRUPO, INGRESO_A_GRUPO, PERIODICO]
    SalidaSnapshot:
      type: object
      additionalProperties: false
      required: [snapshotId]
      properties:
        snapshotId: { type: string, format: uuid }

    # ------------------------------------------------------------------ CU-72
...
    EntradaSellado:
      type: object
      additionalProperties: false
      required: [grupoId, motivo, hechos, desde, hasta, excepcionesDeConciliacionAbiertas]
      properties:
        grupoId: { type: string, format: uuid }
        motivo:  { type: string, enum: [CIERRE_PERIODO, ENTREGA, HITO, DISOLUCION] }
        hechos:  { type: array, items: { $ref: '#/components/schemas/HechoASellar' } }
        desde:   { type: string, format: date-time }
        hasta:   { type: string, format: date-time }
        excepcionesDeConciliacionAbiertas:
          type: integer
          minimum: 0
          description: >-
            Lo resuelve nucleo-financiero y llega resuelto: no se lee su esquema
            (invariante 11) ni se lo consulta dentro de la transaccion (invariante 6).
    SalidaSellado:
      type: object
      additionalProperties: false
      required: [bloqueId, numeroBloque, hashAnterior, raizMerkle, hashBloque, entidadesSelladas]
      properties:
        bloqueId:     { type: string, format: uuid }
        numeroBloque: { type: integer, format: int64 }
        hashAnterior:
          allOf: [{ $ref: '#/components/schemas/Hash' }]
          description: >-
            64 ceros en el primer bloque del grupo. Nunca nulo, porque la columna es
            NOT NULL.
        raizMerkle:   { $ref: '#/components/schemas/Hash' }
        hashBloque:   { $ref: '#/components/schemas/Hash' }
        entidadesSelladas: { type: integer, minimum: 0 }
...
    EntradaEvaluacionInsignias:
      type: object
      additionalProperties: false
      required: [usuarioId, hechos]
      properties:
        usuarioId:       { type: string, format: uuid }
        tipoDeEvento:    { type: [string, 'null'], maxLength: 40 }
        codigosAEvaluar: { type: array, items: { type: string, maxLength: 40 } }
        hechos:          { $ref: '#/components/schemas/HechosDelUsuario' }
    EntradaRevocacionInsignia:
      type: object
      additionalProperties: false
      required: [motivoRevocacion]
      properties:
        motivoRevocacion:
          type: string
          minLength: 10
          maxLength: 160
          description: Obligatorio. Revocar sin motivo deja a la persona sin saber por que.
    SalidaInsignia:
      type: object
      additionalProperties: false
      required: [otorgadaId, otorgadaEn, motivoLegible, revocada, esNueva]
      properties:
        otorgadaId:     { type: string, format: uuid }
        insigniaCodigo: { type: [string, 'null'], maxLength: 40 }
        otorgadaEn:     { type: string, format: date-time }
        motivoLegible:  { type: string }
        revocada:       { type: boolean }
        esNueva:        { type: boolean }

    # ------------------------------------------------------------------ CU-75
    EntradaCertificado:
      type: object
      additionalProperties: false
      required: [usuarioId, snapshotId, disponibles, incluir, vigenciaDias, identidadVerificada]
      properties:
        usuarioId:  { type: string, format: uuid }
        snapshotId: { type: string, format: uuid }
        disponibles:
          type: object
          additionalProperties: { type: string }
          description: Todo lo que se PODRIA mostrar, ya resuelto desde la foto.
        incluir:
          type: array
          items: { type: string }
          description: Lo que el titular eligio. Lo demas no entra ni al hash.
        vigenciaDias:        { type: integer, minimum: 7, maximum: 180, default: 90 }
        identidadVerificada: { type: boolean }
    EntradaRevocacionCertificado:
      type: object
      additionalProperties: false
      required: [motivo]
      properties:
        motivo: { type: string, minLength: 3, maxLength: 300 }
    SalidaCertificado:
      type: object
      additionalProperties: false
      required: [certificadoId, codigoVerificacion, urlPublica, hashContenido, emitidoEn,
                 expiraEn, contenido]
      properties:
        certificadoId:      { type: string, format: uuid }
        codigoVerificacion: { type: string, maxLength: 40 }
        urlPublica:         { type: string }
        hashContenido:      { $ref: '#/components/schemas/Hash' }
        emitidoEn:          { type: string, format: date-time }
        expiraEn:           { type: string, format: date-time }
        contenido:
          type: object
          additionalProperties: { type: string }
...
    EntradaResena:
      type: object
      additionalProperties: false
      required: [grupoId, autorParticipanteId, evaluadoUsuarioId, calificacion, dimension,
                 cicloCerrado, diasDeVentana]
      properties:
        grupoId:             { type: string, format: uuid }
        autorParticipanteId: { type: string, format: uuid }
        evaluadoUsuarioId:   { type: string, format: uuid }
        calificacion:        { type: integer, minimum: 1, maximum: 5 }
        dimension:
          type: string
          description: Catalogo cerrado de ck_resena_participante_dimension.
          enum: [PUNTUALIDAD, COMUNICACION, ORGANIZACION]
        comentario:   { type: [string, 'null'], maxLength: 500 }
        cicloCerrado: { type: boolean }
        cerradoEn:    { type: [string, 'null'], format: date-time }
        diasDeVentana: { type: integer, minimum: 1 }
        autorFueExpulsado:    { type: boolean, default: false }
        resenasPreviasDelAutor: { type: integer, minimum: 0, default: 0 }
    EntradaModeracion:
      type: object
      additionalProperties: false
      required: [decision, motivo]
      properties:
        decision: { type: string, enum: [APROBAR, RECHAZAR, PUBLICAR_SIN_COMENTARIO] }
        motivo:   { type: string, minLength: 10, maxLength: 300 }
    SalidaResena:
      type: object
      additionalProperties: false
      required: [resenaId, estadoModeracion, pesoEnReputacion, marcas]
      properties:
        resenaId: { type: string, format: uuid }
        estadoModeracion:
          type: string
          description: Catalogo cerrado de ck_resena_participante_estado_moderacion.
          enum: [PENDIENTE, PUBLICADA, RECHAZADA, OCULTA]
        pesoEnReputacion:
          type: string
          description: Siempre menor que el peso de los factores de pago.
        motivoModeracion: { type: [string, 'null'] }
        marcas:
          type: array
          items: { type: string, enum: [TELEFONO, CORREO, DOCUMENTO] }
          description: Lo que la primera pasada encontro. Retiene, no recorta.

    # ------------------------------------------------------------------ CU-97
...